Выносит renderTask в класс TaskController, подключает библиотеки flatpickr и moment
